Add: fix import

+++ python-server_side_rendering/task_00_intro.py
@@ -42,6 +42,9 @@ def generate_invitations(template, attendees):
 
         filename = f"output_{index}.txt"
 
+        if os.path.exists(filename):
+            continue
+
         try:
             with open(filename, "w") as file:
                 file.write(output)
